Revert "✨ Feat: 프로필 페이지 최적화 및 팔로우 기능"
