+++ src/components/AddModal/AddModal.style.tsx
@@ -21,7 +21,6 @@ export const Overlay = styled.div`
 `;
 
 export const Wrapper = styled.div<StyleProps>`
-  width: 70%;
   height: ${(props) => props.step === 1 && "70%"}; // 단계에 따라 변경 필요할 듯
   position: absolute;
   background-color: ${COLOR.White};
@@ -99,7 +98,7 @@ export const FirstStepWrapper = styled.div`
   flex-direction: column;
   align-items: center;
   padding: 30px;
-  gap: 15px;
+  gap: 10px;
 `;
 
 export const Span = styled.span`
@@ -134,13 +133,14 @@ export const SecondStepWrapper = styled(FirstStepWrapper)`
 
 // 2단계 이미지 동영상 들어가는 부분
 export const ImgWrapper = styled.div<StyleProps>`
-  /* width: 80%; */
-  flex: 2;
   position: relative;
+  display: flex;
+  align-items: center;
   ${(props) =>
     props.type === "feed"
       ? css`
           width: 500px;
+          height: 500px;
         `
       : css`
           width: 400px;
@@ -148,11 +148,12 @@ export const ImgWrapper = styled.div<StyleProps>`
         `}
 `;
 
-export const Images = styled.div`
-  width: 100%;
+export const Images = styled.div<StyleProps>`
   display: flex;
-  scroll-behavior: smooth;
   overflow: hidden;
+  scroll-behavior: smooth;
+  width: 500px;
+  /* height: 500px; */
 `;
 
 export const ArrowBox = styled.div`
@@ -170,21 +171,18 @@ export const ArrowRightBox = styled(ArrowBox)`
 `;
 // 2단계 글쓰는 부분들
 export const SecondRightWrapper = styled.div`
-  /* width: 340px; */
+  width: 320px;
   display: flex;
-  flex: 1;
   flex-direction: column;
-  height: 100%;
 `;
 
 export const Img = styled.img<StyleProps>`
-  max-width: 100%;
-  object-fit: fill;
   -webkit-user-select: none;
   -khtml-user-select: none;
   -moz-user-select: none;
   -o-user-select: none;
   user-select: none;
+  object-fit: contain;
   ${(props) =>
     props.type === "feed"
       ? css`
@@ -192,7 +190,6 @@ export const Img = styled.img<StyleProps>`
         `
       : css`
           width: 400px;
-          height: 800px;
         `}
 `;
