+++ src/pages/EditProfile/EditProfile.tsx
@@ -25,7 +25,7 @@ function EditProfile() {
     useEditProfileAPI();
   const [editProfileData, setEditProfileData] =
     useRecoilState<EditProfileResponseType>(EditProfileState);
-  const [text, setText] = useState(" ");
+  const [text, setText] = useState("");
   const [isChecked, setIsChecked] = useState(false);
   const [isOn, setIsOn] = useRecoilState<boolean>(ToggleState);
   const [countText, setCountText] = useState(0);
@@ -77,13 +77,7 @@ function EditProfile() {
     //🔥 API
     const requestEdit = async () => {
       try {
-        await requestEditProfile(
-          setEditProfileData,
-          setText,
-          setIsChecked,
-          setIsOn,
-          setCountText,
-        );
+        await requestEditProfile(setEditProfileData, setText, setIsChecked, setIsOn);
         const textlength = text.length;
         setCountText(textlength);
       } catch (error) {
@@ -120,7 +114,7 @@ function EditProfile() {
             <S.EctTitle>소개</S.EctTitle>
             <S.InputBox>
               <S.IntroInput
-                value={text ? text : ""}
+                value={text ? text : editProfileData.userIntro}
                 onChange={(e: React.ChangeEvent<HTMLTextAreaElement>) => {
                   setText(e.target.value);
                   setCountText(e.target.value.length);
