+++ src/pages/SearchTag/SearchTag.style.tsx
@@ -1,11 +1,7 @@
 import styled from "styled-components";
 import * as FONT from "../../constants/font";
 import * as COLOR from "../../constants/color";
-interface PropsType {
-  overlay?: boolean;
-  feedId?: number;
-  ImgId?: number;
-}
+
 export const SearchTagWrapper = styled.div`
   display: flex;
   flex-direction: column;
@@ -42,14 +38,13 @@ export const FeedBox = styled.div`
   justify-content: center;
 `;
 
-export const FeedImg = styled.img<PropsType>`
+export const FeedImg = styled.img`
   width: 100%;
   height: 100%;
   object-fit: cover;
-  filter: ${(PropsType) =>
-    PropsType.overlay && PropsType.feedId === PropsType.ImgId
-      ? "brightness(0.5)"
-      : "brightness(1)"};
+  &:hover {
+    filter: brightness(0.5);
+  }
 `;
 
 export const FeedHoverMutiple = styled.div`
@@ -72,7 +67,7 @@ export const FeedHover = styled.div`
 `;
 
 export const FeedHoverMultiItem = styled.div`
-  padding: 5px 10px;
+  margin: 5px 10px;
   align-self: center;
   justify-content: center;
   font-size: ${FONT.ML};
@@ -81,7 +76,7 @@ export const FeedHoverMultiItem = styled.div`
 `;
 
 export const FeedHoverItem = styled.div`
-  padding: 0px 15px;
+  margin: 0px 15px;
   align-self: center;
   justify-content: center;
   font-size: ${FONT.ML};
